add dummy-colors.ts; move files around and fix typos
