+++ src/util/dummy-colors.ts
@@ -1,0 +1,10 @@
+// These exports mirror the exports in its sibling file, `tty-colors.ts`. It
+// exists to provide compatibility with environments where the `tty` module
+// isn't available.
+
+export const red = String;
+export const green = String;
+export const yellow = String;
+export const blue = String;
+export const magenta = String;
+export const cyan = String;
